// src/types/hr-management/resignation-management.d.ts
export interface IResignationEmployeeProject {
    dcName: string;
    dgName: string;
    effort: number;
    employeeProjectStatus: string;
    startDate: string;
    endDate: string;
    managerName: string;
    positionName: string;
    projectName: string;
}

export interface IResignationEmployeeCommitment {
    commitmentId: number;
    commitmentName: string;
    costFee: number;
    employeeId: number;
    fromDate: string;
    isBroken: boolean;
    notes: string;
    signedDate: string;
    toDate: string;
}

export interface IResignationEmployeeContract {
    contractId: number;
    contractTypeName: string;
    employeeId: number;
    fromDate: string;
    toDate: string;
    signOrder: number;
    isNonTechnical: boolean;
    isRenewalContract: boolean;
    renewApprovalStatusName: string;
}

export interface IResignationEmployeeOnsiteHistory {
    actualEndDate: string;
    cityName: string;
    countryName: string;
    employeeId: number;
    expectedEndDate: string;
    flightDeparture: string;
    flightReturn: string;
    onsiteCountryId: number;
    onsiteFormId: number;
    projectName: string;
    visaTypeName: string;
    customer: string;
}

export interface IResignationEmployeeTemporaryLeave {
    actualEndDate: string;
    effort: number;
    endDate: string;
    projectId: number;
    projectName: string;
    startDate: string;
    statusId: number;
    statusName: string;
    temporaryLeaveId: number;
}

export interface IResignationCancelled {
    projectName: string;
    dgName: string;
    dcName: string;
    applyDate: string;
    cancelledDate: string;
    reasonName: string;
}

export interface IResignationMoreInformation {
    employeeCommitments: IResignationEmployeeCommitment[];
    employeeContracts: IResignationEmployeeContract[];
    employeeOnsiteHistories: IResignationEmployeeOnsiteHistory[];
    employeeProjects: IResignationEmployeeProject[];
    employeeTemporaryLeaves: IResignationEmployeeTemporaryLeave[];
    resignationFormCancelled: IResignationCancelled[];
}

export interface IResignationUnit {
    unitId: number;
    unitName: string;
    parentId: number;
    unitTypeId: number;
    unitTypeName: string;
    isSmallest: boolean;
    children: IResignationUnit[];
}

export interface IResignationIndexes {
    companies: { companyId: number; companyName: string }[];
    positions: { isDefault: boolean; positionId: number; positionName: string }[];
    units: IResignationUnit[];
    reasonForLeave: { reasonId: number; reasonName: string; isAttrition: boolean }[];
    resignationStatuses: { statusId: number; statusName: string; statusTypeId: number; totalEffort: number }[];
    resignationTypes: { resignationTypeId: number; resignationTypeName: string; isOfficialEmployee: boolean }[];
}